feat: move ViewModels and feed states to commons (batch 3)
Move to commons module for KMP compatibility:
- RelayFeedViewModel + MyRelayInfo (pure commons types)
- NewEphemeralChatMetaViewModel (IAccount instead of Account)
- UserFeedState sealed class
- StringFeedState sealed class
- ChannelFeedState sealed class

Also includes IAccountViewModel interface from expand-iaccountviewmodel branch.

Amethyst callers updated with explicit imports from commons.
Build verified with :commons:compileKotlinJvm and :amethyst:compilePlayDebugKotlin.
